refactor(agents): extract lab-handoff reconciliation from lifecycle_ops god file
lifecycle_ops.rs is the single highest-churn file in the control plane (34
commits in the recent window) and a 2685-line god file. Extract the
lab-runner handoff / pending-submission-intent reconciliation subsystem —
the volatile heart that the recent stabilization fixes (#9382, #9396,
#9411, ...) keep touching — into lab_handoff_reconciliation.rs.

Moved: reconcile_active_lab_runner_handoffs, has_/reconcile_pending_
runner_submission_intent, has_accepted_runner_handoff,
has_recorded_provider_progress, run_owes_candidate_follow_up,
bind_pending_runner_submission_if_accepted, expire_unaccepted_lab_handoff,
reconcile_runner_job_state, terminalize_lost_accepted_lab_job.

Sibling module uses 'use super::*' (mod.rs re-exports lifecycle_ops::* and
the sibling modules) and is itself re-exported 'pub use
lab_handoff_reconciliation::*' so the external contract is unchanged. A few
lifecycle_ops items the cluster needs are now pub(crate) (LabHandoffLock +
its lock()); the two callbacks lifecycle_ops still calls
(expire_unaccepted_lab_handoff, reconcile_runner_job_state) are pub(crate);
cancellation.rs's bind_pending import repointed to the new module.

Behavior-preserving: lifecycle_ops.rs 2869 -> 2326 lines,
lab_handoff_reconciliation.rs 554. All 24 handoff_and_proxy tests pass. (3
pre-existing env-dependent flakes — controller_proxy_is_queued_before_
handoff_then_binds_runner_child, pinned_runtime_recovery_retains_the_
existing_lab_proxy_identity, long_pre_submission_setup_survives_
reconciliation — fail identically on clean main; stash-verified.)
