+++ crates/homeboy-agents/src/agent_task_lifecycle/mod.rs
@@ -37,6 +37,7 @@ pub mod controller_pin_reference_provider;
 mod conversion;
 mod failure_recording;
 mod health;
+mod lab_handoff_reconciliation;
 mod lifecycle_candidate_adoption;
 mod lifecycle_ops;
 mod lifecycle_record_ops;
@@ -50,6 +51,7 @@ pub use cancellation::*;
 pub use failure_recording::*;
 pub use health::*;
 pub use homeboy_core::controller_runtime::ControllerRuntimePruneResult;
+pub use lab_handoff_reconciliation::*;
 pub use lifecycle_candidate_adoption::*;
 pub use lifecycle_ops::*;
 pub use lifecycle_record_ops::cook_attempt_run_id;
